Moved persistent file init
