+++ lib/app.js
@@ -72,21 +72,7 @@ wfmSync.init(mediator, mbaasApi, 'result', syncOptions);
 wfmSync.init(mediator, mbaasApi, 'workflows', syncOptions);
 wfmSync.init(mediator, mbaasApi, 'messages', syncOptions);
 
-var mongoUtil = require("./app/mongoUtil");
-mongoUtil.getMongoURL(function(err, url) {
-  if (err || !url) {
-    console.log("Cannot determine valid mongodb connection url. Using temporary file storage");
-    require('fh-wfm-file/lib/cloud')(mediator, app, {usePersistentStorage: false});
-  } else {
-    var storageConfig = {
-      gridFs: {
-        mongoUrl: url
-      }
-    };
-    require('fh-wfm-file-storage/lib/cloud')(mediator, storageConfig);
-    require('fh-wfm-file/lib/cloud')(mediator, app, {usePersistentStorage: true});
-  }
-});
+
 
 require('fh-wfm-message/lib/server')(mediator, app, mbaasApi);
 require('fh-wfm-result/lib/cloud')(mediator);
@@ -105,9 +91,9 @@ module.exports = function() {
     require('./app/result')(mediator),
     require('./app/workflow')(mediator),
     require('./app/group')(mediator),
-    require('./app/file')(mediator)]).then(function() {
+    require('./app/file')(mediator, app)]).then(function() {
     // Important: errorHandler should be the last thing to be added to the express app.
-    app.use(mbaasExpress.errorHandler());
-    return Promise.resolve(app);
-  });
+      app.use(mbaasExpress.errorHandler());
+      return Promise.resolve(app);
+    });
 };
